refactor(ofm_core/contracts): extract renewal cluster; mod.rs is now a shell
Move the 11 remaining renewal-path functions out of contracts/mod.rs into
contracts/renewals.rs: the public entry points evaluate_renewal_offer,
propose_renewal, project_renewal_financial_impact and delegate_renewals, plus the
negotiation-state helpers (renewal_outcome, next_renewal_round,
cool_stale_renewal_session, build_renewal_feedback, renewal_blocked_until,
should_manual_renewal_fail_on_relationship, has_active_manager_block).

This is the final cluster of the contracts.rs decomposition. mod.rs 706 -> 157 lines
and is now a thin shell: crate imports, module wiring/re-exports, the public API types
(RenewalOffer/Outcome/Decision, RenewalFinancialProjection, the DelegatedRenewal* and
ContractTermination* structs, ContractWarningStage) and the shared ContractReleaseReason
enum. The renamed `_service` imports carry into renewals.rs via `use super::*`.

Byte-faithful move via script; four public entry points re-exported via `pub use`.

Summary of the full decomposition (contracts.rs 1475 -> mod.rs 157 + 6 submodules):
consts 17, helpers 268, free_agent 213, termination 154, expiry 152, renewals 560.

387 ofm_core lib + 28 contracts integration tests pass; clippy clean in contracts/.
